Change: Move withFoldToggle HOC into an own module
