+++ src/web/components/folding/__tests__/Folding.test.tsx
@@ -4,13 +4,8 @@
  */
 
 import {describe, test, expect} from '@gsa/testing';
-import {render, screen, fireEvent} from 'web/testing';
-import {
-  withFolding,
-  FoldState,
-  withFoldToggle,
-  type FoldStateType,
-} from 'web/components/folding/Folding';
+import {render, screen} from 'web/testing';
+import {withFolding, FoldState} from 'web/components/folding/Folding';
 
 describe('withFolding', () => {
   const DummyComponent = (props: React.HTMLAttributes<HTMLDivElement>) => (
@@ -26,34 +21,5 @@ describe('withFolding', () => {
 
     rerender(<FoldableComponent foldState={FoldState.UNFOLDED} />);
     expect(screen.getByText('Dummy Component')).toBeVisible();
-  });
-});
-
-describe('withFoldToggle', () => {
-  test('toggles foldState when onFolded isCalled', () => {
-    interface DummyProps {
-      foldState: FoldStateType;
-      onFoldToggle: () => void;
-    }
-
-    const DummyComponent = ({foldState, onFoldToggle}: DummyProps) => (
-      <div>
-        <span data-testid="foldState">{foldState}</span>
-        <button onClick={onFoldToggle}>Toggle</button>
-      </div>
-    );
-
-    const FoldToggleComponent = withFoldToggle(DummyComponent);
-
-    render(<FoldToggleComponent />);
-
-    expect(screen.getByTestId('foldState')).toHaveTextContent(
-      FoldState.UNFOLDED,
-    );
-
-    fireEvent.click(screen.getByText('Toggle'));
-    expect(screen.getByTestId('foldState')).toHaveTextContent(
-      FoldState.FOLDING_START,
-    );
   });
 });
